+++ packages/react/jest.config.js
@@ -1,6 +1,9 @@
 module.exports = {
     testPathIgnorePatterns : ['node_modules'],
     moduleFileExtensions : ['ts', 'tsx', 'js'],
+    setupFiles: [
+        './jestSetup.js'
+    ],
     transform : {
         '^.+\\.(ts|tsx)$' : 'ts-jest'
     },
