+++ packages/react/jestSetup.js
@@ -1,0 +1,4 @@
+const { configure } = require('enzyme');
+const Adapter = require('enzyme-adapter-react-16');
+
+configure({ adapter: new Adapter() });
