+++ packages/react/package.json
@@ -22,11 +22,11 @@
     "build": "tsc"
   },
   "dependencies": {
-    "@bem/sdk.naming.entity.stringify": "1.0.0"
+    "@bem/sdk.naming.entity.stringify": "1.0.3"
   },
   "devDependencies": {
     "@bem/sdk.entity-name": "0.2.6",
-    "@bem/sdk.naming.presets": "0.0.9",
+    "@bem/sdk.naming.presets": "0.0.10",
     "@types/enzyme": "3.1.10",
     "@types/jest": "22.2.3",
     "@types/react": "16.3.14",
