+++ packages/react/test/content.ts
@@ -1,43 +1,37 @@
+import { createElement } from 'react';
+import { Bem, Block, Elem, withMods } from '../src';
 import { getNode } from './helpers/node';
-import * as BemReact from './helpers/react';
-import { run } from './helpers/run';
 
-type Preset = typeof BemReact /*| BemPreact*/;
+describe('Content:', () => {
+    describe('Bem:', () => {
+        it('has passed content', () => {
+            expect(getNode(createElement(Bem, {
+                block: 'Block',
+                children: 'content'
+            })).text()).toBe('content');
+        });
+    });
 
-run({ BemReact }, (preset: Preset) => () => {
-    const { Bem, Block, render } = preset;
-
-    describe('Content:', () => {
-        describe('Bem:', () => {
-            it('has passed content', () => {
-                expect(getNode(render(Bem, {
-                    block: 'Block',
-                    children: 'content'
-                })).text()).toBe('content');
-            });
+    describe('Block', () => {
+        it('has passed content', () => {
+            class MyBlock extends Block {
+                protected block = 'MyBlock';
+            }
+            expect(getNode(createElement(MyBlock, {
+                children: 'content'
+            })).text()).toBe('content');
         });
 
-        describe('Block', () => {
-            it('has passed content', () => {
-                class MyBlock extends Block {
-                    protected block = 'MyBlock';
+        it('has declared and passed content', () => {
+            class MyBlock extends Block {
+                protected block = 'MyBlock';
+                protected content() {
+                    return this.props.children + 'declared';
                 }
-                expect(getNode(render(MyBlock, {
-                    children: 'content'
-                })).text()).toBe('content');
-            });
-
-            it('has declared and passed content', () => {
-                class MyBlock extends Block {
-                    protected block = 'MyBlock';
-                    protected content() {
-                        return this.props.children + 'declared';
-                    }
-                }
-                expect(getNode(render(MyBlock, {
-                    children: 'content'
-                })).text()).toBe('contentdeclared');
-            });
+            }
+            expect(getNode(createElement(MyBlock, {
+                children: 'content'
+            })).text()).toBe('contentdeclared');
         });
     });
 });
