+++ packages/react/test/displayName.ts
@@ -1,37 +1,31 @@
+import { createElement } from 'react';
+import { Bem, Block, Elem, withMods } from '../src';
 import { getMountedNode } from './helpers/node';
-import * as BemReact from './helpers/react';
-import { run } from './helpers/run';
 
-type Preset = typeof BemReact /*| BemPreact*/;
+describe('DisplayName:', () => {
 
-run({ BemReact }, (preset: Preset) => () => {
-    const { Bem, Block, Elem, render } = preset;
-
-    describe('DisplayName:', () => {
-
-        describe('Bem:', () => {
-            it('generates name in runtime', () => {
-                expect(getMountedNode(render(Bem, { block: 'Block' })).name()).toBe('Bem');
-            });
+    describe('Bem:', () => {
+        it('generates name in runtime', () => {
+            expect(getMountedNode(createElement(Bem, { block: 'Block' })).name()).toBe('Bem');
         });
+    });
 
-        describe('Block:', () => {
-            it('generates name in runtime', () => {
-                class MyBlock extends Block {
-                    protected block = 'MyBlock';
-                }
-                expect(getMountedNode(render(MyBlock)).name()).toBe('MyBlock');
-            });
+    describe('Block:', () => {
+        it('generates name in runtime', () => {
+            class MyBlock extends Block {
+                protected block = 'MyBlock';
+            }
+            expect(getMountedNode(createElement(MyBlock)).name()).toBe('MyBlock');
         });
+    });
 
-        describe('Elem:', () => {
-            it('generates name in runtime', () => {
-                class MyElem extends Elem {
-                    protected block = 'MyBlock';
-                    protected elem = 'Shalala';
-                }
-                expect(getMountedNode(render(MyElem)).name()).toBe('MyBlock-Shalala');
-            });
+    describe('Elem:', () => {
+        it('generates name in runtime', () => {
+            class MyElem extends Elem {
+                protected block = 'MyBlock';
+                protected elem = 'Shalala';
+            }
+            expect(getMountedNode(createElement(MyElem)).name()).toBe('MyBlock-Shalala');
         });
     });
 });
