--- packages/react/test/helpers/run.ts
@@ -1,5 +1,0 @@
-export const run = (libs: { [key: string]: any }, spec: any) => {
-    Object.keys(libs).forEach((libName) => {
-        describe(`${libName.replace('Bem', '')}:`, spec(libs[libName]));
-    });
-};
