+++ packages/react/tsconfig.json
@@ -1,31 +1,32 @@
 {
     "compilerOptions": {
+        "declaration": true,
         "target": "es5",
-		"module": "commonjs",
-		"moduleResolution": "node",
-		"lib": [
+        "module": "commonjs",
+        "moduleResolution": "node",
+        "lib": [
             "dom",
             "es7"
         ],
 
-		"rootDir": "src",
-		"outDir": "lib",
+        "rootDir": "src",
+        "outDir": "lib",
 
-		"strictFunctionTypes": true,
-		"strictNullChecks": true,
+        "strictFunctionTypes": true,
+        "strictNullChecks": true,
 
-		"noImplicitAny": true,
-		"noImplicitReturns": true,
-		"noImplicitThis": true,
-		"noFallthroughCasesInSwitch": true,
-		"noUnusedLocals": true,
-		"noUnusedParameters": true,
+        "noImplicitAny": true,
+        "noImplicitReturns": true,
+        "noImplicitThis": true,
+        "noFallthroughCasesInSwitch": true,
+        "noUnusedLocals": true,
+        "noUnusedParameters": true,
 
-		"downlevelIteration": true,
+        "downlevelIteration": true,
 
-		"pretty": true
-	},
-	"include": [
-		"src/*"
-	]
+        "pretty": true
+    },
+    "include": [
+        "src/*"
+    ]
 }
